+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 ## v1.7.0
 
-*2025-02-10*
+*2025-02-11*
 
 **Added**
 
 - Auto-linking for URLs in Wikitext
-- Hover information for magic words in Wikitext
+- Hover information and signature help for magic words in Wikitext
 
 **Fixed**